Reorganize integration test to allow more with different starting positions
