+++ moveit_ros/moveit_servo/CMakeLists.txt
@@ -130,7 +130,7 @@ if(BUILD_TESTING)
   # End servo parameters launch test
 
   # Servo integration launch test
-  ament_add_gtest_executable(test_servo_integration test/servo_cpp_interface_test.cpp test/test_parameter_struct.hpp)
+  ament_add_gtest_executable(test_servo_integration test/servo_cpp_interface_test.cpp test/test_parameter_struct.hpp test/servo_launch_test_common.hpp)
   ament_target_dependencies(test_servo_integration ${THIS_PACKAGE_INCLUDE_DEPENDS})
   add_ros_test(test/launch/test_servo_integration.test.py ARGS "test_binary_dir:=${CMAKE_CURRENT_BINARY_DIR}")
   # end servo integration launch test
