+++ moveit_ros/moveit_servo/test/config/start_positions.yaml
@@ -1,0 +1,19 @@
+fake_joint_driver_node:
+  ros__parameters:
+    start_position:
+      joints:
+        - panda_joint1
+        - panda_joint2
+        - panda_joint3
+        - panda_joint4
+        - panda_joint5
+        - panda_joint6
+        - panda_joint7
+      values:
+        - 0.0
+        - -0.785
+        - 0.0
+        - -2.356
+        - 0.0
+        - 1.571
+        - 0.785
